# tools/sparkbox/internal/restapi/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "sparkbox control API",
    "version": "1.0.0",
    "summary": "Everything `ssh ctl@<domain>` does, over HTTPS.",
...
    "contact": {
      "name": "sparkbox",
      "url": "https://example.com"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "servers": [
    {
      "url": "https://api.example.com",
      "description": "This host."
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "sessionCookie": []
    }
  ],
...
  "paths": {
...
    "/v1/repos": {
      "get": {
        "operationId": "repo.list",
        "tags": ["repos"],
        "summary": "List the repos attached to your tags",
        "description": "Configuration, not state: a row here says a repository *should* be in every sandbox carrying one of its tags, not that any particular sandbox has finished cloning it. Retagging a running sandbox never clones into it — the guest's `sparkbox repos` shows the new entry as not cloned, and `sparkbox repos sync` is the user's own decision.",
        "responses": {
          "200": {
            "description": "Your attachments.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RepoList" },
                "example": {
                  "repos": [
                    {
                      "id": "9f8b1c2d3e4f5a6b",
                      "owner": "vanpelt",
                      "host": "github.com",
                      "slug": "wandb/hivemind",
                      "ref": "main",
                      "access": "read",
                      "tags": ["hm"],
                      "created_at": "2026-08-01T09:00:00Z"
                    }
                  ]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "501": { "$ref": "#/components/responses/Disabled" }
        }
      },
      "post": {
        "operationId": "repo.add",
        "tags": ["repos"],
        "summary": "Attach a repo to your tags",
        "description": "Attaching is a piece of configuration; it clones nothing right now. Sandboxes created afterwards with a matching tag check the repo out at boot.\n\nAn empty `tags` takes the `default` tag, which is also stamped on every untagged sandbox — so an untagged attachment reaches every box you create from then on. The answer names the tags it actually used.\n\nWhen this host has a GitHub App configured, the call also asks github.com whether the App can reach the repository and answers with `check`. That probe is advisory: the attachment is saved either way, because the App is usually installed a minute later by somebody who is not you, and refusing the write would leave you with nothing recorded and an errand to run.\n\nAttaching requires a GitHub link proved by a published key or the device flow — a login you merely typed is not evidence anyone may mint an installation token for you. No link at all is the 409; a link proved by somebody else's assertion is the 403.",
        "parameters": [{ "$ref": "#/components/parameters/IdempotencyKey" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RepoRequest" },
              "example": {
                "slug": "wandb/hivemind",
                "tags": ["hm"],
                "ref": "main"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The attachment, the tags it took, and whether the App can reach it.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RepoResult" },
                "example": {
                  "repo": {
                    "id": "9f8b1c2d3e4f5a6b",
                    "owner": "vanpelt",
                    "host": "github.com",
                    "slug": "wandb/hivemind",
                    "ref": "main",
                    "access": "read",
                    "tags": ["hm"],
                    "created_at": "2026-08-01T09:00:00Z"
                  },
                  "sandboxes": ["demo"],
                  "defaulted": false,
                  "check": {
                    "host": "github.com",
                    "slug": "wandb/hivemind",
                    "tags": ["hm"],
                    "access": "read",
                    "checked": true,
                    "reachable": false,
                    "reason": "the GitHub App is not installed on that repository",
                    "install_url": "https://github.com/apps/sparkbox/installations/new"
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Invalid" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": {
            "description": "Your GitHub link was not proved directly with GitHub, so it cannot attach repositories. Re-link with `POST /v1/account/github` or the device flow.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "409": {
            "description": "No GitHub account is linked to this handle, so there is nobody to clone as.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "501": { "$ref": "#/components/responses/Disabled" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "spk_v1.<claims>.<hmac>",
        "description": "An edge session token. Mint one with `ssh ctl@example.com session-token | tr -d '\\r\\n'` (the channel ends the line CRLF, and a stray carriage return in the header is rejected), or `POST /v1/account/tokens` from an existing session."
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "spark_session",
        "description": "The same token, as the zone-wide cookie a browser gets from `https://login.example.com`. State-changing calls made with the cookie alone must also prove first-party intent — an `Origin` matching this host, or the header `X-Sparkbox-Console: 1` — because the cookie rides every same-site request. Bearer-authenticated calls need neither."
      }
    },
    "parameters": {
...
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": false,
        "description": "An opaque key of your choosing. The first response is remembered for 24 hours and replayed verbatim (with `Idempotency-Replayed: true`) if the key comes back. Reusing a key with a different request body is a 422; sending it while the first request is still running is a 409. Server errors are never cached, so a 5xx can always be retried with the same key.",
        "schema": { "type": "string" },
        "example": "3f1c9e2a-7b64-4c0e-9a3b-5d2f8c1e6a04"
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "No valid session. Send `Authorization: Bearer <token>`.",
        "content": {
          "text/plain": {
            "schema": { "type": "string" }
          }
        }
      },
...
      "Invalid": {
        "description": "The arguments are malformed.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": {
              "error": {
                "kind": "invalid",
                "op": "resize",
                "code": "bad_size",
                "message": "bad size \"25T\" — use e.g. 25G or 512M"
              }
            }
          }
        }
      },
      "Disabled": {
        "description": "This host has not configured that feature. Check `GET /v1/capabilities` first.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": {
              "error": {
                "kind": "disabled",
                "op": "archive",
                "code": "archive_disabled",
                "message": "archiving is not enabled on this host"
              }
            }
          }
        }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "description": "Every failure has this shape. Switch on `error.code`.",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["kind", "op", "code", "message"],
            "properties": {
              "kind": {
                "type": "string",
                "description": "The broad class of failure, which fixes the status code.",
                "enum": [
                  "internal",
                  "invalid",
                  "not_found",
                  "denied",
                  "conflict",
                  "disabled",
                  "limit",
                  "capacity",
                  "quota",
                  "upstream"
                ]
              },
              "op": {
                "type": "string",
                "description": "The operation that failed — the same string as the operationId.",
                "example": "pause"
              },
              "code": {
                "type": "string",
                "description": "A stable machine token. Never localized, never changed once shipped.",
                "example": "sandbox_not_found"
              },
              "message": {
                "type": "string",
                "description": "A complete sentence, fit to show a user unedited."
              },
              "hint": {
                "type": "string",
                "description": "What to do about it, when there is something to do."
              },
              "details": {
                "type": "object",
                "description": "Structured facts a client can act on: `running` for a limit, `budget_mb` for capacity, `matches` for an ambiguous passkey prefix.",
                "additionalProperties": true
              }
            }
          }
        }
      },
...
      "RepoList": {
        "type": "object",
        "required": ["repos"],
        "properties": {
          "repos": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Repo" }
          }
        }
      },
      "RepoResult": {
        "type": "object",
        "description": "The stored attachment plus the three things you cannot see from the request: which of your sandboxes now select it, whether it landed on the `default` tag, and what github.com says about it.",
        "required": ["repo", "sandboxes", "defaulted", "check"],
        "properties": {
          "repo": { "$ref": "#/components/schemas/Repo" },
          "sandboxes": {
            "type": "array",
            "items": { "type": "string" },
            "description": "Your existing sandboxes carrying one of the tags. Nothing is cloned into them by this call — retagging never clones — but an attachment that reaches nothing is the mistake worth catching immediately.",
            "example": ["demo"]
          },
          "defaulted": {
            "type": "boolean",
            "description": "The request named no tag, so the attachment went on `default` — which every untagged sandbox also carries, meaning this repo now clones into every box you create. Say so to whoever asked."
          },
          "check": { "$ref": "#/components/schemas/RepoCheck" }
        }
      },
      "RepoCheck": {
        "type": "object",
        "description": "One attachment's answer to \"can the GitHub App actually reach this\". `checked` and `reachable` are two bits because they fail differently: `checked` false means this host holds no App at all, so only public repositories will clone and the attachment itself is fine; `reachable` false means there is an App and it cannot get at this repository, which is yours to fix.",
        "required": ["host", "slug", "access", "tags", "checked", "reachable"],
        "properties": {
          "host": { "type": "string", "example": "github.com" },
          "slug": { "type": "string", "example": "wandb/hivemind" },
          "tags": {
            "type": "array",
            "items": { "type": "string" },
            "example": ["hm"]
          },
          "access": { "type": "string", "enum": ["read", "write"] },
          "checked": { "type": "boolean", "description": "This host has a GitHub App, so the answer below means something." },
          "reachable": { "type": "boolean", "description": "The App is installed on the repository and your GitHub identity may use that installation." },
          "reason": {
            "type": "string",
            "description": "Why not, as a sentence fit to show a user. Absent when reachable.",
            "example": "the GitHub App is not installed on that repository"
          },
          "install_url": {
            "type": "string",
            "description": "Present when installing the App is the fix.",
            "example": "https://github.com/apps/sparkbox/installations/new"
          }
        }
      },
...
      "RepoRequest": {
        "type": "object",
        "required": ["slug"],
        "properties": {
          "slug": {
            "type": "string",
            "description": "`owner/name`, with no `.git` suffix and no URL around it.",
            "example": "wandb/hivemind"
          },
          "host": {
            "type": "string",
            "description": "Absent means github.com, which is the only value this host serves.",
            "example": "github.com"
          },
          "tags": {
            "type": "array",
            "items": { "type": "string" },
            "description": "The whole tag set. Empty or absent takes `default`, which every untagged sandbox also carries — so the repo lands in every box you create from then on, and the answer sets `defaulted`.",
            "example": ["hm"]
          },
          "ref": {
            "type": "string",
            "description": "Branch or tag. Absent takes the repository's default branch.",
            "example": "main"
          },
          "path": {
            "type": "string",
            "description": "Clone destination, relative to the login user's home. Absent takes the default layout.",
            "example": "src/hm"
          },
          "write": {
            "type": "boolean",
            "description": "Ask for a token that can push. Absent means read-only, which is what an agent that only reads code should get. The stored row reports this back as `access`."
          }
        }
      },
...
    }
  }
}